Vector-DB/requirements.txt
langchain-pinecone==0.2.3
langchain-openai==0.0.8
langchain-community==0.0.27
langchain-core==0.1.32
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
chromadb==0.4.24
sentence-transformers==2.5.1
streamlit==1.32.2
python-dotenv==1.0.1
numpy==1.26.4